Changed namespace and package name to Microsoft.AspNetCore.Razor.TagHelpers.Test
