+++ src/Microsoft.AspNetCore.Razor.TagHelpers.Test/CaseSensitiveTagHelperDescriptorComparer.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Razor.Compilation.TagHelpers;
 using Microsoft.Extensions.Internal;
 using Xunit;
 
-namespace Microsoft.AspNetCore.Razor.Test.Internal
+namespace Microsoft.AspNetCore.Razor.TagHelpers.Test
 {
     internal class CaseSensitiveTagHelperDescriptorComparer : TagHelperDescriptorComparer
     {
