+++ src/Microsoft.AspNetCore.Razor.TagHelpers.Test/TagHelperDesignTimeDescriptorComparer.cs
@@ -7,21 +7,18 @@ using Microsoft.AspNetCore.Razor.Compilation.TagHelpers;
 using Microsoft.Extensions.Internal;
 using Xunit;
 
-namespace Microsoft.AspNetCore.Razor.Test.Internal
+namespace Microsoft.AspNetCore.Razor.TagHelpers.Test
 {
-    internal class TagHelperAttributeDesignTimeDescriptorComparer :
-        IEqualityComparer<TagHelperAttributeDesignTimeDescriptor>
+    internal class TagHelperDesignTimeDescriptorComparer : IEqualityComparer<TagHelperDesignTimeDescriptor>
     {
-        public static readonly TagHelperAttributeDesignTimeDescriptorComparer Default =
-            new TagHelperAttributeDesignTimeDescriptorComparer();
+        public static readonly TagHelperDesignTimeDescriptorComparer Default =
+            new TagHelperDesignTimeDescriptorComparer();
 
-        private TagHelperAttributeDesignTimeDescriptorComparer()
+        private TagHelperDesignTimeDescriptorComparer()
         {
         }
 
-        public bool Equals(
-            TagHelperAttributeDesignTimeDescriptor descriptorX,
-            TagHelperAttributeDesignTimeDescriptor descriptorY)
+        public bool Equals(TagHelperDesignTimeDescriptor descriptorX, TagHelperDesignTimeDescriptor descriptorY)
         {
             if (descriptorX == descriptorY)
             {
@@ -32,15 +29,18 @@ namespace Microsoft.AspNetCore.Razor.Test.Internal
             Assert.NotNull(descriptorY);
             Assert.Equal(descriptorX.Summary, descriptorY.Summary, StringComparer.Ordinal);
             Assert.Equal(descriptorX.Remarks, descriptorY.Remarks, StringComparer.Ordinal);
+            Assert.Equal(descriptorX.OutputElementHint, descriptorY.OutputElementHint, StringComparer.Ordinal);
 
             return true;
         }
 
-        public int GetHashCode(TagHelperAttributeDesignTimeDescriptor descriptor)
+        public int GetHashCode(TagHelperDesignTimeDescriptor descriptor)
         {
             var hashCodeCombiner = HashCodeCombiner.Start();
+
             hashCodeCombiner.Add(descriptor.Summary, StringComparer.Ordinal);
             hashCodeCombiner.Add(descriptor.Remarks, StringComparer.Ordinal);
+            hashCodeCombiner.Add(descriptor.OutputElementHint, StringComparer.Ordinal);
 
             return hashCodeCombiner;
         }
