+++ src/Microsoft.AspNetCore.Razor.TagHelpers.Test/project.json
