+++ test/Microsoft.AspNetCore.Razor.Runtime.Test/TagHelpers/TagHelperAttributeListTest.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.AspNetCore.Razor.Test.Internal;
+using Microsoft.AspNetCore.Razor.TagHelpers.Test;
 using Xunit;
 
 namespace Microsoft.AspNetCore.Razor.TagHelpers
