+++ test/Microsoft.AspNetCore.Razor.Runtime.Test/project.json
@@ -7,7 +7,7 @@
   "dependencies": {
     "dotnet-test-xunit": "2.2.0-*",
     "Microsoft.AspNetCore.Razor.Runtime": "1.1.0-*",
-    "Microsoft.AspNetCore.Razor.Test.Sources": {
+    "Microsoft.AspNetCore.Razor.TagHelpers.Test": {
       "version": "1.1.0-*",
       "type": "build"
     },
