+++ test/Microsoft.AspNetCore.Razor.Test/project.json
@@ -4,7 +4,7 @@
     "dotnet-test-xunit": "2.2.0-*",
     "Microsoft.AspNetCore.Razor": "1.1.0-*",
     "Microsoft.AspNetCore.Razor.Runtime": "1.1.0-*",
-    "Microsoft.AspNetCore.Razor.Test.Sources": {
+    "Microsoft.AspNetCore.Razor.TagHelpers.Test": {
       "version": "1.1.0-*",
       "type": "build"
     },
